KM-013: adding artist model to the admin dashboard and listing display and search fields
