+++ music/models.py
@@ -1,11 +1,11 @@
 # -*- coding: utf-8 -*-
 import uuid
 from django.db import models
-from account.models import State
+from account.models import State, Artist
 
 
 class Album(models.Model):
-    artist = models.CharField(max_length=60)
+    artist = models.ForeignKey(Artist, on_delete=models.CASCADE)
     album_title = models.CharField(max_length=30, unique = True)
     genre = models.CharField(max_length=20, null = True, default = 'Other')
     state = models.ForeignKey(State, on_delete=models.CASCADE)
@@ -30,15 +30,3 @@ class Song(models.Model):
     
     def __str__(self):
         return '%s%s' % (self.song_title, self.file_type)
-
-
-class Artist(models.Model):
-    name = models.CharField(max_length = 50)
-    alias = models.CharField(max_length = 50)
-    state = models.ForeignKey(State, on_delete=models.CASCADE)
-    key = models.CharField(max_length = 255, default = uuid.uuid4, editable = False)
-    date_modified = models.DateTimeField(auto_now = True)
-    date_created = models.DateTimeField(auto_now_add = True)
-
-    def __str__(self):
-        return self.name
